Made navbar a common component
