+++ parks/src/buyersPage/index.js
@@ -1,9 +1,20 @@
-import React, { useEffect, lazy, Suspense } from "react";
+import React, { lazy, Suspense } from "react";
+
+const NavbarComponent = lazy(() => import("../common/navbar"));
 
 const BuyersPage = () => {
+  const renderNavbar = () => {
+    return (
+      <Suspense fallback={""}>
+        <NavbarComponent />
+      </Suspense>
+    );
+  };
+
   return (
     <div>
-      <h3>Buyers Page</h3>
+      {renderNavbar()}
+      <h3 style={{ marginTop: 60 }}>Buyers Page</h3>
     </div>
   );
 };
