+++ parks/src/common/styles.js
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import { Navbar } from "react-bootstrap";
+
+export const NavbarHome = styled(Navbar)`
+  position: fixed;
+  left: 0;
+  right: 0;
+  top: 0;
+`;
