+++ parks/src/home/index.js
@@ -1,11 +1,11 @@
-import React, { useEffect, lazy, Suspense } from "react";
+import React, { lazy, Suspense } from "react";
 // import GoogleLogin from "react-google-login";
 import { useNavigate } from "react-router-dom";
 
-
 // Style imports
-import { HomeContainer, AvtarBox, AvtarContainer, NavbarHome } from "./styles";
-import { Navbar, Nav, Container } from "react-bootstrap";
+import { HomeContainer, AvtarBox, AvtarContainer } from "./styles";
+
+const NavbarComponent = lazy(() => import("../common/navbar"));
 
 const Home = () => {
   let navigate = useNavigate();
@@ -20,13 +20,20 @@ const Home = () => {
   //   dispatch(newUserRegistraion());
   // }, [dispatch]);
 
-  const responseGoogle = (response) => {
-    console.log(response);
-    console.log(response.profileObj);
+  // const responseGoogle = (response) => {
+  //   console.log(response);
+  //   console.log(response.profileObj);
+  // };
+
+  const renderNavbar = () => {
+    return (
+      <Suspense fallback={""}>
+        <NavbarComponent />
+      </Suspense>
+    );
   };
 
   const handleOnBuyersClick = () => {
-    // history.push("/buyer");
     navigate("/buyer");
   };
 
@@ -34,27 +41,9 @@ const Home = () => {
     navigate("/renter");
   };
 
-  const handleOnLoginPage = () => {
-    navigate("/login");
-  };
-  const handleOnRegisterClick = () => {
-    navigate("/registration");
-  };
-  const handleOnAbout = () => {
-    navigate("/about");
-  };
   return (
     <HomeContainer>
-      <NavbarHome bg="dark" variant="dark">
-        <Container>
-          <Navbar.Brand href="#home">ParkS</Navbar.Brand>
-          <Nav className="me-auto">
-            <Nav.Link onClick={handleOnAbout}>About</Nav.Link>
-            <Nav.Link onClick={handleOnLoginPage}>Login</Nav.Link>
-            <Nav.Link onClick={handleOnRegisterClick}>Register</Nav.Link>
-          </Nav>
-        </Container>
-      </NavbarHome>
+      {renderNavbar()}
 
       <AvtarContainer>
         <AvtarBox onClick={handleOnRentersClick}>
