+++ parks/src/login/styles.js
@@ -1,8 +1,5 @@
 import styled from "styled-components";
-import homeBg from "../assets/Parks.jpeg";
-import { Navbar } from "react-bootstrap";
 
-import { NavbarHome } from "../home/styles";
 export const LoginContainer = styled.div`
   display: flex;
   align-items: center;
